lagere extractRelevantFolder aus
